requirements.txt
annotated-types==0.7.0
astroid==3.3.9
autoflake==2.3.1
bcrypt==4.3.0
black==25.1.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_cytoscape==1.0.2
dill==0.4.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.15.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
plotly==5.24.1
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
pylint==3.3.6
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
retrying==1.3.4
setuptools==75.3.0
six==1.16.0
tenacity==9.0.0
tomlkit==0.13.2
types-click==7.1.8
types-Flask==1.1.6
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
types-Werkzeug==1.0.9
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
